requirements.txt
discord.py
python-dotenv
asyncio
yt-dlp
pynacl